Beginning on History reader. Got testing to work.
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3885 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/org/mpisws/p2p/transport/peerreview/history/SecureHistory.java
@@ -41,8 +41,8 @@ import java.nio.ByteBuffer;
 
 import rice.p2p.util.RandomAccessFileIOBuffer;
 
-public interface SecureHistory {
-  public int getNumEntries();
+public interface SecureHistory /* extends Iterable<IndexEntry> */ {
+  public long getNumEntries();
   public long getBaseSeq();
   public long getLastSeq();
